requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
opencv-python>=4.8.1.78
opencv-contrib-python>=4.8.1.78
Pillow>=10.3.0
pillow-heif>=0.13.0
pillow-avif-plugin>=1.4.3
pdf2image>=1.16.3
numpy>=1.26.0
scikit-image>=0.22.0
ultralytics>=8.0.196
torch>=2.1.0
torchvision>=0.16.0
efficientnet-pytorch>=0.7.1
scikit-learn>=1.3.2
paddleocr>=2.7.0.3
paddlepaddle>=2.5.2
# detectron2>=0.6  # Optional - Windows compatibility issues, install via WSL or alternative method
fvcore>=0.1.5
iopath>=0.1.7
# OCR Pipeline Dependencies (SERVICE 3)
transformers>=4.30.0  # Required for TrOCR recognition engine
loguru>=0.7.2
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2










